lvgui: Move general stuff in __init
